% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{IA_cols}
\alias{IA_cols}
\title{Function to extract IA_colors as hex codes}
\usage{
IA_cols(...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
returns the vector colors defined in the palette.
}
\details{
currently there are 8 colors.
}
\examples{
IA_cols()[1] is the black color. IA_cols()[2] is the gray.
}
