% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_plot.R
\name{IA_density1}
\alias{IA_density1}
\title{Density plot for one variable with mean line and text}
\usage{
IA_density1(
  data,
  var,
  fill_color = 3,
  mean_line = TRUE,
  line_color = 2,
  text = FALSE
)
}
\arguments{
\item{data}{a data.frame or a tibble or anything that fits into ggplot}

\item{var}{is a continues variable to plot its density}

\item{fill_color}{fill color from values 1 to 8 taken from the IA palette}

\item{mean_line}{add a mean line True or False}

\item{line_color}{line color from values 1 to 8 taken from the IA palette}

\item{text}{add text labels with the mean and standard deviation}
}
\value{

}
\description{
Density plot for one variable with mean line and text
}
\examples{
IA_density1(mtcars, qsec, fill_color = 3, mean_line = T, text = T)

}
