% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{IA_remove_y}
\alias{IA_remove_y}
\title{remove the y axis.}
\usage{
IA_remove_y()
}
\value{

}
\description{
the line, ticks, text and title. just add it to a ggplot function as the last theme adjustment.
}
\examples{
ggplot(mpg, aes(x = displ, y = cty, color = drv)) +
geom_point() + IA_remove_y()
}
