% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{IA_rotate_x_text}
\alias{IA_rotate_x_text}
\title{Rotate the x axis text}
\usage{
IA_rotate_x_text(angle = 90, hjust = NULL, vjust = NULL, ...)
}
\arguments{
\item{angle}{numeric value specifying the rotation angle. Default is 90 for vertical x-axis text.}

\item{hjust}{horizontal adjustments}

\item{vjust}{vertical adjustments}

\item{...}{other arguments to pass to the function element_text().}
}
\value{

}
\description{
Rotate the x axis text
}
\examples{
add it to a ggplot.
}
