% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{IA_theme}
\alias{IA_theme}
\title{theme for use in Israeli Antitrust Authority.}
\usage{
IA_theme(base_size = 12, base_family = "Alef")
}
\arguments{
\item{base_size}{specify your font size}

\item{base_family}{specify font name}
}
\value{

}
\description{
can be added in a ggplot command of set as default theme with "theme_set(theme_IA())".
Before use it for the first time make sure the font "Alef" exist in the systme and that R
can see it. use extrafont::font_import to import system fonts into R.
This should take a few minutes.
}
\examples{
ggplot(mpg, aes(x = displ, y = cty, color = drv)) +
geom_point(size = 4) + IA_theme() +
labs(title = "type in Hebrew and make shure the font is Alef")



}
