% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/home_page.R
\docType{package}
\name{IAplots - general instructions}
\alias{IAplots - general instructions}
\title{IAplots A package for plotting with Israeli Competition Authority colors}
\description{
The IAplots package privids scale colors and theme for use in the organiztion plotting. Also,
there are a few helper functions to prevent pasting the same code over again
}
\section{the functions a based on ggplot2}{
 to use the the theme thare's also
a  need for the package extrafont. for the theme to load the Alef font,
make sure the Alef is imported into R. loading the fonts takes a while, but then
one can define any font wanted to his plots.
}

