% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{Scale_Color_and_Fill}
\alias{Scale_Color_and_Fill}
\alias{IA_scale_color}
\alias{IA_scale_fill}
\title{Scale color and fill}
\usage{
IA_scale_color(palette = "DBP", discrete = TRUE, reverse = FALSE, ...)

IA_scale_fill(palette = "DBP", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{chose palette from a list of options.
currently there are 13 variants. the colors are marked with capital letters.
"D" - Dark (black), "B" Blue, "G" gray and P pink.
the options are: DP, DB, DG, GB, GP, BP,
DBP, DGP, DGB, GBP, c4, c5, c8.}

\item{discrete}{specify if the color/fill variable is discrete or continoues}

\item{reverse}{reverse colors of the palette.}

\item{...}{pass other arguments to the  scale parameters such as "guide = "none""
or "position = "left",see ggplot discrete_scale for more options.}
}
\value{

}
\description{
Add scale_color or scale_fill to a ggplot object with IA color palettes of choice
}
\examples{
just add scale_color_IA at any position inside a ggplot command:

 ggplot(mpg, aes(x = displ, y = cty, color = drv)) +
 geom_point(size = 4) +
 IA_scale_color(palette = "GBP", reverse = T)
}
